- Tested OrientDB module succesfully - Changed actor to not transient - Can't return unmodifiable event list because of OrientDB

+++ trunk/40_implementation/LIMO/LIMO_OrientDB/src/test/java/nl/fontys/sofa/limo/orientdb/mock/MockOrientDBAccess.java
@@ -15,12 +15,16 @@ public class MockOrientDBAccess extends OrientDBAccess {
 
     @Override
     public void closeConnection() {
+        checkConnection();
+        
         if (connection != null) {
-            if (connection.isClosed()) {
-                connection.open("admin", "admin");
+            if (connection.exists()) {
+                if (connection.isClosed()) {
+                    connection.open("admin", "admin");
+                }
+                
+                connection.drop();
             }
-            
-            connection.drop();
             super.closeConnection();
         }
     }
